requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
chardet==5.2.0
faiss-cpu==1.7.4
langchain==0.0.281
langsmith==0.0.92
openai==0.28.0
pandas==2.2.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
python-telegram-bot==21.3
python_docx==1.1.2
Requests==2.32.3
scikit_learn==1.5.0
setuptools==63.2.0
tiktoken==0.5.1
uvicorn==0.30.1
